Prueba3
El personaje se mueve en las 4 direcciones de manera correcta??
