+++ Assets/EscenaPruebas.unity
@@ -185,13 +185,13 @@ Transform:
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 125773583}
-  m_LocalRotation: {x: 0.27059805, y: 0.6532815, z: -0.27059805, w: 0.6532815}
-  m_LocalPosition: {x: -4.26, y: 4.13, z: -0.7}
+  m_LocalRotation: {x: 0.42261827, y: 0, z: 0, w: 0.9063079}
+  m_LocalPosition: {x: 0, y: 4.66, z: -5.57}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 45.000004, y: 90, z: 0}
+  m_LocalEulerAnglesHint: {x: 50, y: 0, z: 0}
 --- !u!1 &700147461
 GameObject:
   m_ObjectHideFlags: 0
@@ -383,23 +383,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 1636469429}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6ef3b339bf2389f46aaaa115b19a58eb, type: 3}
+  m_Script: {fileID: 11500000, guid: 963f9a6a01c705742900aca711acf134, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  jumpImpulse: 0
-  maxMoveSpeed: 10
+  VelocidadMax: 5
   controller: 0
-  matRed: {fileID: 0}
-  matGreen: {fileID: 0}
-  matBlue: {fileID: 0}
-  matYellow: {fileID: 0}
-  laserAPrefab: {fileID: 0}
-  laserBPrefab: {fileID: 0}
-  laserXPrefab: {fileID: 0}
-  laserYPrefab: {fileID: 0}
-  laserBumpPrefab: {fileID: 0}
-  triggerLeftPrefab: {fileID: 0}
-  triggerRightPrefab: {fileID: 0}
 --- !u!114 &1636469435
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -411,6 +399,19 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c255eeabbc5f89940a4bf2628352cca0, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &1636469436
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1636469429}
+  m_Enabled: 0
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f42b4b7688bda784991bb9852fd6a93f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  horizontalSpeed: 2
+  verticalSpeed: 2
 --- !u!1001 &2126469710
 Prefab:
   m_ObjectHideFlags: 0
