--- Assets/Scripts/MoveLaser.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b7128a2731b06dc40a7d2b8896566d04
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
