--- Assets/Scripts/MovePlayer.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6ef3b339bf2389f46aaaa115b19a58eb
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
