+++ Assets/Scripts/Transicion.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: f42b4b7688bda784991bb9852fd6a93f
+timeCreated: 1508079191
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
